+++ src/hooks/init/window.js
@@ -27,6 +27,7 @@ import { generateIcon } from "../../lib/icons.js";
 import DDBProxy from "../../lib/DDBProxy.js";
 import DDBCompanion from "../../parser/companions/DDBCompanion.js";
 import DDBCompanionFactory from "../../parser/companions/DDBCompanionFactory.js";
+import { configureCustomAAForCondition, addSaveAdvantageToTarget } from "../../effects/helpers.js";
 
 function resetSecrets() {
   game.settings.set("ddb-importer", "cobalt-cookie-local", false);
@@ -99,5 +100,9 @@ export function registerWindow() {
     characterSettings: MuncherSettings.getCharacterImportSettings,
     DDBCompanion,
     DDBCompanionFactory,
+    effects: {
+      configureCustomAAForCondition,
+      addSaveAdvantageToTarget,
+    }
   };
 }